Remove propagated Tag LastUpdate
Some resource, like CloudFormation Distribution, take time to be updated.
Does it make sense to have a tag with LastUpdateTime even if resource properties are not changed at all ?
If a resource is created by CloudFormation i can simply look at Stack LastUpdateTime
to have the same information derived by tagging it (i know that with tagging is simpler to do this).