can enforce StackType when using "-t" appending "+"
Normally when doing stack selection matching StackType, a stack is
selected if any stack type found after "-t" is found inside the Tag
StackType.
For StackType "ec2 ecs" they are selected if is used:
-t ecs
or
-t ec2
And i can select ONLY "ec2 ecs" using:
-t "ec2 ecs"

But what if i want to apply a change ONLY to ecs or ec2 and do not
include "ec2 ecs" too ????

Now it can be accomplished appending a plus sign ("+") to the stack type
(must be a single one).

Ex:
-t ecs+
or
-t ec2+