ChangeSet Simplification
by default simply ChangeSet:
looking at Replacement filed Create a list of resources that do not need to be replaced.
Then for every resource in the ChangeSet compare his CausingEntity list,
if present, with the list of not changed resources.
If all CausingEntity items are related to resources not changed, skip
that resource in the ChangeSet output.
(i do not even know why AWS not directly do this!!!)

Using parameter --changeset-original will revert the behaviour and will show
Original not simplyfied ChangeSet.