Add --tags options to create/update
can be used to add custom tags to be assigned to the stack.
They will be propagated to supported instances.
Tags need to be space separated and a "=" is used to separate the Tag
Name from the Tag Value.
Ex:
--tags price_zone=test bill_company=hello_world

To remove an already existing tag, use "REMOVE" as Tag Value