Changeset simplification fixes - changeset output
During changeset simplification i need to consider that if i modify the
template and some resource property value change, they will be listed in
the target field but there will not be a relative item in
CausingEntityList.
So if i take in account only CausingEntityList i can skip a resource
that will be changed by a change in the template.
To avoid this i add another condition before removing a changeset row:
Targets items but be less that CausingEntityList ones.
(there can be a problem if one resource property is "defined" by two
Parameters or, even less probable, by two other resources; but this is
unlikely to occur)

I have changed the ouput of Target and CausingEntity.
Now items are not splitted by newlines but by comma.
This way table output is more readable.