Tags, Metadata and Cmd update or add
When processing Metadata and Command line Tags i need to take in account
if Tag Key already exist but with different value.
In that case i need to update the current one and not append another one
(Remenber that Tag is a list of dict).

Created helper function add_or_update_tag to do this.